[FIX] for bug introduced in r10657
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/branches/Release1.10@10660 6adb6e08-d915-0410-941f-83917bcadc18

+++ source/APPLICATIONS/TOPP/OpenSwathConfidenceScoring.C
@@ -312,7 +312,7 @@ protected:
     OpenSwath::Scoring::normalize_sum(&assay_intensities[0],
                                       boost::numeric_cast<int>(assay_intensities.size()));
     DoubleReal dist_int = manhattanDist_(feature_intensities, 
-                                         boost::numeric_cast<int>(assay_intensities));
+                                         assay_intensities);
 
     DoubleReal score = glm_(diff_rt, dist_int);
